requirements.txt
django
django-rest-framework
drf_spectacular
Pillow